refactor(macos): split privileged helper XPC owners
