+++ clients/apple/scripts/architecture-warning-baseline.txt
@@ -1,3 +1,3 @@
 # Exact Apple architecture-warning debt. Keep entries sorted.
 # large entries include their current line ceiling; every reduction tightens it.
-large|Services/Shell/AlanPrivilegedHelperXPC.swift|2117
+# No accepted warnings remain.
